support all blackmagic modes
modify setup to use getDisplayMode, uses int for framerate
shouldn’t break existing code as compatible types.

framerates can be short (29) or verbose (2997)

+++ src/ofxBlackMagic.h
@@ -23,7 +23,7 @@ private:
 	
 public:
 	ofxBlackMagic();
-	bool setup(int width, int height, float framerate);
+	bool setup(int width, int height, int framerate);
 	void close(); // should call this in ofApp::exit()
 	bool update(); // returns true if there is a new frame
